% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-surveys.R
\name{apply_gear_mapping_v1}
\alias{apply_gear_mapping_v1}
\title{Gear mapping for version 1 data}
\usage{
apply_gear_mapping_v1()
}
\description{
Gear mapping for version 1 data
}
\keyword{internal}
