% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-surveys.R
\name{apply_gear_mapping_v2}
\alias{apply_gear_mapping_v2}
\title{Gear mapping for version 2 data}
\usage{
apply_gear_mapping_v2()
}
\description{
Gear mapping for version 2 data
}
\keyword{internal}
