% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracks.R
\name{calculate_fishing_summaries}
\alias{calculate_fishing_summaries}
\title{Calculate fishing effort metrics}
\usage{
calculate_fishing_summaries(gps_data, interval_seconds = 300)
}
\arguments{
\item{gps_data}{SF object with classified activities}

\item{interval_seconds}{Time interval between positions}
}
\value{
List with effort summaries
}
\description{
Summarizes fishing effort by trip and spatial grid
}
\keyword{internal}
