% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{check_outliers_iqr}
\alias{check_outliers_iqr}
\title{Check for outliers using IQR method}
\usage{
check_outliers_iqr(x, multiplier = 1.5)
}
\arguments{
\item{x}{numeric vector where outliers will be checked}

\item{multiplier}{multiplier for IQR range (default is 1.5)}
}
\value{
a logical vector indicating which values are within bounds (TRUE) or outliers (FALSE)
}
\description{
Check for outliers using IQR method
}
\examples{
\dontrun{
x <- c(1, 2, 3, 100)
check_outliers_iqr(x, multiplier = 1.5)
}
}
\keyword{validation}
