% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-surveys.R
\name{clean_catch_names}
\alias{clean_catch_names}
\title{Clean Catch Names}
\usage{
clean_catch_names(data = NULL)
}
\arguments{
\item{data}{A data frame or tibble containing a column named \code{catch_name}.}
}
\value{
A data frame or tibble with standardized catch names in the \code{catch_name} column.
}
\description{
This function standardizes catch names in the dataset by correcting common misspellings
and inconsistencies.
}
\examples{
\dontrun{
cleaned_data <- clean_catch_names(data)
}

}
\keyword{preprocessing}
