% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracks.R
\name{create_exclusion_zones}
\alias{create_exclusion_zones}
\title{Create port exclusion zones}
\usage{
create_exclusion_zones(ports_df, buffer_m = 500)
}
\arguments{
\item{ports_df}{Data frame with port locations}

\item{buffer_m}{Buffer radius in meters}
}
\value{
SF object with unified exclusion zone
}
\description{
Creates buffer zones around ports to exclude anchored vessels
}
