% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracks.R
\name{create_extent_polygon}
\alias{create_extent_polygon}
\title{Create study area extent}
\usage{
create_extent_polygon(gps_data)
}
\arguments{
\item{gps_data}{SF object with GPS positions}
}
\value{
SF polygon defining study extent
}
\description{
Creates bounding polygon for study area
}
\keyword{internal}
