% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{get_catch_bounds_iqr}
\alias{get_catch_bounds_iqr}
\title{Get fish groups Catch Bounds using IQR method}
\usage{
get_catch_bounds_iqr(data = NULL, multiplier = 1.5)
}
\arguments{
\item{data}{A data frame containing columns: gear, fish_category, catch_kg}

\item{multiplier}{multiplier for IQR range (default is 1.5)}
}
\value{
A data frame with columns: gear, fish_category, upper.up
}
\description{
Get fish groups Catch Bounds using IQR method
}
