% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{get_total_catch_bounds}
\alias{get_total_catch_bounds}
\title{Get Total Catch Bounds}
\usage{
get_total_catch_bounds(data = NULL, k = NULL)
}
\arguments{
\item{data}{A data frame containing columns: \code{gear} and \code{total_catch_kg}.}

\item{k}{A numeric value used in the \code{univOutl::LocScaleB} function for outlier detection.}
}
\value{
A data frame with columns: \code{gear} and \code{upper.up} (the upper bound).
}
\description{
Calculates the upper bounds for \emph{total} catch data (using \code{total_catch_kg})
based on gear type.
}
