% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{ingest_landings}
\alias{ingest_landings}
\title{Download and Process WCS Catch Surveys from Kobotoolbox}
\usage{
ingest_landings(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{Logging threshold level (default: logger::DEBUG)}
}
\value{
No return value. Function downloads data, processes it, and uploads to Google Cloud Storage.
}
\description{
This function retrieves survey data from Kobotoolbox for a specific project,
processes it, and uploads the raw data as a Parquet file to Google Cloud Storage.
It uses the \code{get_kobo_data} function, which is a wrapper for \code{kobotools_kpi_data}
from the KoboconnectR package.
}
\details{
The function performs the following steps:
\enumerate{
\item Reads configuration settings.
\item Downloads survey data from Kobotoolbox using \code{get_kobo_data}.
\item Checks for uniqueness of submissions.
\item Converts data to tabular format.
\item Uploads raw data as a Parquet file to Google Cloud Storage.
}

Note that while parameters are provided for customization, the function
currently uses hardcoded values and configuration settings for some parameters.
}
\examples{
\dontrun{
ingest_landings()
}
}
\keyword{ingestion}
\keyword{workflow}
