% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{ingest_surveys}
\alias{ingest_surveys}
\title{Download WCS Surveys from Kobotoolbox}
\usage{
ingest_surveys(
  url = NULL,
  project_id = NULL,
  username = NULL,
  psswd = NULL,
  encoding = NULL
)
}
\arguments{
\item{url}{The URL of kobotoolbox (often referred to as 'kpi-url').}

\item{project_id}{Is the asset id of the asset for which the data is
to be downloaded.}

\item{username}{Username of your kobotoolbox account.}

\item{psswd}{Password of the account.}

\item{encoding}{Encoding to be used. Default is "UTF-8".}
}
\value{
A character vector with paths of the downloaded files.
}
\description{
This function retrieves survey data from Kobotoolbox for a specific project.
It allows users to customize the filename using a prefix, choose between CSV or RDS formats,
and decide whether to append versioning information to the filename.
The resulting files are downloaded to the working directory or specified path,
with paths returned as a character vector.
}
\examples{
\dontrun{
file_list <- retrieve_wcs_surveys(
  url = "kf.kobotoolbox.org",
  project_id = "my_project_id",
  username = "admin",
  psswd = "admin",
  encoding = "UTF-8"
)
}
}
