% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracks.R
\name{prepare_gps_data}
\alias{prepare_gps_data}
\title{Prepare GPS data in GPSMonitoring format}
\usage{
prepare_gps_data(tracks)
}
\arguments{
\item{tracks}{Data frame with vessel tracking data}
}
\value{
SF object with formatted GPS data
}
\description{
Converts raw tracking data into format required by GPSMonitoring package
}
