% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-surveys.R
\name{preprocess_landings_core}
\alias{preprocess_landings_core}
\title{Core preprocessing logic for landings data}
\usage{
preprocess_landings_core(raw_dat, gear_mapping_func)
}
\arguments{
\item{raw_dat}{Raw data from cloud storage}

\item{gear_mapping_func}{Function that handles gear standardization}
}
\value{
Preprocessed landings data
}
\description{
Core preprocessing logic for landings data
}
\keyword{internal}
