% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracks.R
\name{process_trajectories_with_speed}
\alias{process_trajectories_with_speed}
\title{Calculate vessel speeds from GPS positions}
\usage{
process_trajectories_with_speed(
  gps_data,
  trip_gap_seconds = 7200,
  standardize = TRUE,
  interval_seconds = 300
)
}
\arguments{
\item{gps_data}{SF object with GPS data}

\item{trip_gap_seconds}{Seconds between pings to define new trip}

\item{standardize}{Whether to standardize time intervals}

\item{interval_seconds}{Target interval in seconds if standardizing}
}
\value{
SF object with speed calculations
}
\description{
Processes trajectories and calculates speed between consecutive points
}
\keyword{internal}
