% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_legacy_landings}
\alias{validate_legacy_landings}
\title{Validate Legacy Fisheries Data}
\usage{
validate_legacy_landings()
}
\value{
This function does not return a value. Instead, it processes the data and uploads the validated results to a MongoDB collection in the pipeline databse.
}
\description{
This function imports and validates preprocessed legacy fisheries data from a MongoDB collection. It conducts a series of validation checks to ensure data integrity, including checks on dates, fisher counts, boat numbers, and catch weights. The function then compiles the validated data and corresponding alert flags, which are subsequently uploaded back to MongoDB.
}
\details{
The function performs the following main operations:
\enumerate{
\item Pulls preprocessed legacy landings data from the preprocessed MongoDB collection.
\item Validates the data for consistency and accuracy, focusing on:
\itemize{
\item Date validation
\item Number of fishers
\item Number of boats
\item Catch weight
}
\item Generates a validated dataset that integrates the results of the validation checks.
\item Creates alert flags to identify and track any data issues discovered during validation.
\item Merges the validated data with additional metadata, such as survey details and landing site information.
\item Uploads the validated dataset to the validated MongoDB collection.
}
}
\note{
This function requires a configuration file to be present and readable by the 'read_config' function, which should provide MongoDB connection details and parameters for validation.
}
\examples{
\dontrun{
validate_legacy_landings()
}

}
