% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_nboats}
\alias{validate_nboats}
\title{Validate Number of Boats}
\usage{
validate_nboats(data = NULL)
}
\arguments{
\item{data}{A data frame containing the \code{n_boats} column.}
}
\value{
A data frame with two columns: \code{n_boats} and \code{alert_n_boats}.
\itemize{
\item \code{n_boats}: The original number of boats if valid, otherwise \code{NA}.
\item \code{alert_n_boats}: A numeric value indicating the alert (error label) number.
}
}
\description{
This function validates the \code{n_boats} column in the provided dataset.
An alert (error label) is triggered if the number of boats is an outlier,
determined by the \code{alert_outlier} function with specified parameters. The alert number
is stored in \code{alert_n_boats}. If an alert is triggered, the \code{n_boats} value
is set to \code{NA}.
}
\examples{
\dontrun{
validate_nboats(data)
}
}
