% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_nboats_iqr}
\alias{validate_nboats_iqr}
\title{Validate Number of Boats using IQR method}
\usage{
validate_nboats_iqr(data = NULL, multiplier = 1.5, flag_value = NULL)
}
\arguments{
\item{data}{A data frame containing the n_boats column}

\item{multiplier}{multiplier for IQR range (default is 1.5)}

\item{flag_value}{A numeric value to use as the flag for values outside bounds}
}
\value{
A data frame with validated n_boats and alert flags
}
\description{
Validate Number of Boats using IQR method
}
\examples{
\dontrun{
validate_nboats_iqr(data, multiplier = 1.5, flag_value = 8)
}
}
