% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_nfishers}
\alias{validate_nfishers}
\title{Validate Number of Fishers}
\usage{
validate_nfishers(data = NULL, k = NULL, flag_value = NULL)
}
\arguments{
\item{data}{A data frame containing the \code{no_of_fishers} column.}

\item{k}{a numeric value used in the LocScaleB function for outlier detection.}

\item{flag_value}{A numeric value to use as the flag for catches exceeding the upper bound.}
}
\value{
A data frame with two columns: \code{no_of_fishers} and \code{alert_n_fishers}.
\itemize{
\item \code{no_of_fishers}: The original number of fishers if valid, otherwise \code{NA}.
\item \code{alert_n_fishers}: A numeric value indicating the alert (error label) number.
}
}
\description{
This function validates the \code{no_of_fishers} column in the provided dataset.
An alert (error label) is triggered if the number of fishers is an outlier,
determined by the \code{alert_outlier} function with specified parameters. The alert number
is stored in \code{alert_n_fishers}. If an alert is triggered, the \code{no_of_fishers} value
is set to \code{NA}.
}
\examples{
\dontrun{
validate_nfishers(data, k = 3, flag_value = 2)
}
}
\keyword{validation}
