% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_nfishers_iqr}
\alias{validate_nfishers_iqr}
\title{Validate Number of Fishers using IQR method}
\usage{
validate_nfishers_iqr(data = NULL, multiplier = 1.5, flag_value = NULL)
}
\arguments{
\item{data}{A data frame containing the no_of_fishers column}

\item{multiplier}{multiplier for IQR range (default is 1.5)}

\item{flag_value}{A numeric value to use as the flag for values outside bounds}
}
\value{
A data frame with validated no_of_fishers and alert flags
}
\description{
Validate Number of Fishers using IQR method
}
\examples{
\dontrun{
validate_nfishers_iqr(data, multiplier = 1.5, flag_value = 7)
}
}
\keyword{validation}
