% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_summaries}
\alias{export_summaries}
\title{Export Summarized Fishery Data}
\usage{
export_summaries()
}
\value{
This function does not return a value. It processes the data and uploads
the result to MongoDB.
}
\description{
This function reshapes validated legacy fishery data, calculates summary metrics
such as effort and CPUE (Catch Per Unit Effort), and uploads the processed data
to MongoDB for dashboard usage.
}
\details{
The function performs the following operations:
\enumerate{
\item Pulls validated data from the "legacy-validated" MongoDB collection.
\item Reshapes the data by grouping by survey_id and summarizing catch and price.
\item Joins with BMU (Beach Management Unit) size data.
\item Calculates effort and CPUE.
\item Selects relevant columns for the final dataset.
\item Uploads the processed data to a MongoDB collection.
}

Effort and CPUE are calculated as follows:
\itemize{
\item Effort = Number of fishers / Size of BMU in km
\item CPUE = Total catch in kg / Effort
}
}
\note{
This function requires a configuration file to be present and readable by the
'read_config' function, which should provide MongoDB connection details.
It also assumes the existence of a 'bmu_size' dataset for joining.
}
\examples{
\dontrun{
export_summaries()
}

}
\keyword{export}
\keyword{workflow}
