% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{get_catch_bounds}
\alias{get_catch_bounds}
\title{Get fish groups Catch Bounds}
\usage{
get_catch_bounds(data = NULL, k = NULL)
}
\arguments{
\item{data}{A data frame containing columns: \code{gear}, \code{fish_category}, \code{catch_kg}.}

\item{k}{A numeric value used in the \code{univOutl::LocScaleB} function for outlier detection.}
}
\value{
A data frame with columns: \code{gear}, \code{fish_category}, and \code{upper.up} (the upper bound).
}
\description{
Calculates the upper bounds for \emph{fish groups} catch data (using \code{catch_kg})
based on gear type and fish category. Data is grouped by the interaction of gear
and fish category, and category "0" is excluded from the analysis.
}
\keyword{validation}
