% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-landings.R
\name{preprocess_legacy_landings}
\alias{preprocess_legacy_landings}
\title{Preprocess Legacy Landings Data}
\usage{
preprocess_legacy_landings(landings = NULL)
}
\arguments{
\item{landings}{A data frame containing legacy WCS SSF landings data. The data frame
should include the columns: \code{landing_date}, \code{catch_name}, \code{gear}, \code{gear_new},
\code{fish_category}, and \code{ecology}.}
}
\value{
A tibble with transformed data:
\item{landing_date}{Converted to \code{Date} format.}
\item{catch_name, gear, gear_new, fish_category, ecology}{Converted to lowercase.}
}
\description{
This function preprocesses legacy landings data by converting it into a tibble,
transforming the \code{landing_date} column to \code{Date} format, and converting
specified columns to lowercase.
}
\examples{
\dontrun{
preprocessed_data <- preprocess_legacy_landings(landings_data)
}

}
